Remove more Shared aliases
