--- crates/fluvio-spu/src/core/replica/metadata.rs
@@ -1,21 +1,0 @@
-use fluvio_controlplane_metadata::partition::Replica;
-use fluvio_controlplane_metadata::partition::ReplicaKey;
-
-use crate::core::Spec;
-use crate::core::LocalStore;
-
-impl Spec for Replica {
-    const LABEL: &'static str = "Replica";
-
-    type Key = ReplicaKey;
-
-    fn key(&self) -> &Self::Key {
-        &self.id
-    }
-
-    fn key_owned(&self) -> Self::Key {
-        self.id.clone()
-    }
-}
-
-pub type ReplicaStore = LocalStore<Replica>;
