+++ crates/fluvio-spu/src/replication/leader/leaders_state.rs
@@ -1,7 +1,5 @@
-use std::{
-    ops::{Deref},
-};
-use std::sync::RwLock;
+use std::ops::Deref;
+use std::sync::{RwLock, Arc};
 use std::collections::HashMap;
 
 use tracing::{error, instrument};
@@ -9,16 +7,12 @@ use tracing::{error, instrument};
 use fluvio_controlplane_metadata::partition::{Replica, ReplicaKey};
 use fluvio_storage::{FileReplica, StorageError};
 
-use crate::{
-    control_plane::SharedStatusUpdate,
-    core::{GlobalContext},
-};
+use crate::core::GlobalContext;
 use crate::config::ReplicationConfig;
 use crate::replication::follower::FollowerReplicaState;
 
 use super::{LeaderReplicaState, replica_state::SharedLeaderState};
-
-pub type SharedReplicaLeadersState<S> = ReplicaLeadersState<S>;
+use crate::control_plane::StatusMessageSink;
 
 /// Collection of replicas
 #[derive(Debug)]
@@ -39,7 +33,7 @@ impl<S> Deref for ReplicaLeadersState<S> {
 }
 
 impl<S> ReplicaLeadersState<S> {
-    pub fn new_shared() -> SharedReplicaLeadersState<S> {
+    pub fn new_shared() -> ReplicaLeadersState<S> {
         Self::default()
     }
 }
@@ -76,7 +70,7 @@ impl ReplicaLeadersState<FileReplica> {
         &self,
         ctx: &GlobalContext<FileReplica>,
         replica: Replica,
-        status_update: SharedStatusUpdate,
+        status_update: Arc<StatusMessageSink>,
     ) -> Result<LeaderReplicaState<FileReplica>, StorageError> {
         let replica_id = replica.id.clone();
 
@@ -114,7 +108,7 @@ impl ReplicaLeadersState<FileReplica> {
         config: ReplicationConfig,
         follower: FollowerReplicaState<FileReplica>,
         replica: Replica,
-        status_update: SharedStatusUpdate,
+        status_update: Arc<StatusMessageSink>,
     ) -> LeaderReplicaState<FileReplica> {
         let replica_id = replica.id.clone();
         let replica_storage = follower.inner_owned();
